Timer fixing timings

+++ Firmware/channels.h
@@ -13,16 +13,16 @@
 
 void CHANNELS_Configuration(void);
 
-void ch1_puls(uint16_t puls);
-void ch2_puls(uint16_t puls);
-void ch3_puls(uint16_t puls);
-void ch4_puls(uint16_t puls);
-void ch5_puls(uint16_t puls);
-void ch6_puls(uint16_t puls);
-void ch7_puls(uint16_t puls);
-void ch8_puls(uint16_t puls);
+void ch1_puls(int puls);
+void ch2_puls(int puls);
+void ch3_puls(int puls);
+void ch4_puls(int puls);
+void ch5_puls(int puls);
+void ch6_puls(int puls);
+void ch7_puls(int puls);
+void ch8_puls(int puls);
 
-void cppm_output(uint16_t ch1, uint16_t ch2, uint16_t ch3, uint16_t ch4, uint16_t ch5, uint16_t ch6, uint16_t ch7, uint16_t ch8 );
+void cppm_output(int ch1, int ch2, int ch3, int ch4, int ch5, int ch6, int ch7, int ch8 );
 
 
 #endif
