+++ Firmware/esp8266.c
@@ -3,14 +3,14 @@
 #include <stdio.h>
 #include <string.h>
 
-uint16_t ch1=1000;
-uint16_t ch2=1000;
-uint16_t ch3=1000;
-uint16_t ch4=1000;
-uint16_t ch5=1000;
-uint16_t ch6=1000;
-uint16_t ch7=1000;
-uint16_t ch8=1000;
+int ch1=1000;
+int ch2=1000;
+int ch3=1000;
+int ch4=1000;
+int ch5=1000;
+int ch6=1000;
+int ch7=1000;
+int ch8=1000;
 
 void WIFI_Init(void)
 {
